Modify anomaly detection
Timeout countdown will begin as soon as the environment is ready, rather than after the player signaled to start.

The ball's position is checked if it leaves the game environment. It was occasionally escaping through the boundaries in some cases.

In rare cases the ball would halt or slow excessively.